Separated Dataset, Seed. Removed BinaryMask.
