+++ scilpy/image/datasets.py
@@ -1,15 +1,16 @@
 # -*- coding: utf-8 -*-
 import numpy as np
+
 from dipy.core.interpolation import trilinear_interpolate4d, \
-    nearestneighbor_interpolate
+                                    nearestneighbor_interpolate
 
 
-class Dataset(object):
+class AccessibleVolume(object):
     """
     Class to access/interpolate data from nibabel object
     """
 
-    def __init__(self, img, interpolation=None):
+    def __init__(self, img, interpolation=None, must_be_3d=False):
         """
         Parameters
         ----------
@@ -19,6 +20,8 @@ class Dataset(object):
             The interpolation choice amongst "trilinear" or "nearest". If
             None, functions getting a coordinate in mm instead of voxel
             coordinates are not available.
+        must_be_3d: bool
+            If True, dataset can't be 4D.
         """
         self.interpolation = interpolation
         if self.interpolation:
@@ -29,6 +32,10 @@ class Dataset(object):
 
         self.pixdim = img.header.get_zooms()[:3]
         self.data = img.get_fdata(caching='unchanged', dtype=np.float64)
+
+        if must_be_3d and self.data.ndim != 3:
+            raise Exception("Data should have been 3D but data dimension is:"
+                            "{}".format(self.data.ndim))
 
         # Expand dimensionality to support uniform 4d interpolation
         if self.data.ndim == 3:
@@ -187,136 +194,3 @@ class Dataset(object):
             True if position is in dataset range and false otherwise.
         """
         return self.is_voxel_in_bound(*self.get_voxel_at_position(x, y, z))
-
-
-class SeedGenerator(Dataset):
-    """
-    Class to get seeding positions
-    """
-
-    def __init__(self, img):
-        super(SeedGenerator, self).__init__(img, None)
-        self.seeds = np.array(np.where(np.squeeze(self.data) > 0)).transpose()
-
-    def get_next_pos(self, random_generator, indices, which_seed):
-        """
-        Generate the next seed position.
-
-        Parameters
-        ----------
-        random_generator : numpy random generator
-            Initialized numpy number generator.
-        indices : List
-            Indices of current seeding map.
-        which_seed : int
-            Seed number to be processed.
-
-        Return
-        ------
-        seed_pos: tuple
-            Position of next seed expressed in mm.
-        """
-        len_seeds = len(self.seeds)
-        if len_seeds == 0:
-            return []
-
-        half_voxel_range = [self.pixdim[0] / 2,
-                            self.pixdim[1] / 2,
-                            self.pixdim[2] / 2]
-
-        # Voxel selection from the seeding mask
-        ind = which_seed % len_seeds
-        x, y, z = self.seeds[indices[np.asscalar(ind)]]
-
-        # Subvoxel initial positioning
-        r_x = random_generator.uniform(-half_voxel_range[0],
-                                       half_voxel_range[0])
-        r_y = random_generator.uniform(-half_voxel_range[1],
-                                       half_voxel_range[1])
-        r_z = random_generator.uniform(-half_voxel_range[2],
-                                       half_voxel_range[2])
-
-        return x * self.pixdim[0] + r_x, y * self.pixdim[1] \
-            + r_y, z * self.pixdim[2] + r_z
-
-    def init_pos(self, random_initial_value, first_seed_of_chunk):
-        """
-        Initialize numpy number generator according to user's parameter
-        and indexes from the seeding map.
-
-        Parameters
-        ----------
-        random_initial_value : int
-            The "seed" for the random generator.
-        first_seed_of_chunk : int
-            Number of seeds to skip (skip parameter + multi-processor skip).
-
-        Return
-        ------
-        random_generator : numpy random generator
-            Initialized numpy number generator.
-        indices : List
-            Indices of current seeding map.
-        """
-        random_generator = np.random.RandomState(random_initial_value)
-        indices = np.arange(len(self.seeds))
-        random_generator.shuffle(indices)
-
-        # Skip to the first seed of the current process' chunk,
-        # multiply by 3 for x,y,z
-        # Divide the generation to prevent RAM overuse
-        seed_to_go = np.asscalar(first_seed_of_chunk) * 3
-        while seed_to_go > 100000:
-            random_generator.rand(100000)
-            seed_to_go -= 100000
-        random_generator.rand(seed_to_go)
-
-        return random_generator, indices
-
-
-class BinaryMask(object):
-    """
-    Mask class for binary mask.
-    """
-
-    def __init__(self, tracking_dataset):
-        self.m = tracking_dataset
-        # force memmap to array. needed for multiprocessing
-        self.m.data = np.array(self.m.data)
-        ndim = self.m.data.ndim
-        if not (ndim == 3 or (ndim == 4 and self.m.data.shape[-1] == 1)):
-            raise ValueError('mask cannot be more than 3d')
-
-    def isPropagationContinues(self, pos):
-        """
-        The propagation continues if the position is within the mask.
-
-        Parameters
-        ----------
-        pos : tuple
-            3D positions.
-
-        Return
-        ------
-        value: bool
-            True if the position is inside the mask.
-        """
-        return (self.m.get_position_value(*pos) > 0
-                and self.m.is_position_in_bound(*pos))
-
-    def isStreamlineIncluded(self, pos):
-        """
-        If the propagation stoped, this function determines if the streamline
-        is included in the tractogram. Always True for BinaryMask.
-
-        Parameters
-        ----------
-        pos : tuple
-            3D positions.
-
-        Return
-        ------
-        value: bool
-            Always True.
-        """
-        return True
